+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## staged
 
-- none.
+- Added support for multinetwork ITD power flows using `solve_mn_pfitd(...)`.
 
 ## v0.9.0